requirements.txt
fastapi
uvicorn
streamlit
transformers
torch
faiss-cpu
sentence-transformers
PyPDF2
